refactor: extract PostPreview drag and virtual post logic (#186)
Extract useSkipSlotConfirmation, usePostDropOnVirtualSlot hooks and
PostPreviewCard subcomponent from 358-line PostPreview. No behavior
changes.
